Moves ParserDispatcher tests back to parse.py, adds some stubs
