+++ provenance_lib/tests/test_parse.py
@@ -10,7 +10,10 @@ from networkx import DiGraph
 from networkx.classes.reportviews import NodeView  # type: ignore
 
 from ..checksum_validator import ChecksumDiff, ValidationCode
-from ..parse import ProvDAG, UnparseableDataError, ProvDAGParser
+from ..parse import (
+    ProvDAG, UnparseableDataError, ProvDAGParser, ParserDispatcher,
+)
+from ..util import UUID
 from ..zipfile_parser import (
     ParserV0, ParserV1, ParserV2, ParserV3, ParserV4, ParserV5,
     Config, ProvNode, ParserResults,
@@ -965,3 +968,64 @@ class ProvDAGParserTests(unittest.TestCase):
             self.assertEqual(parsed.provenance_is_valid,
                              dag.provenance_is_valid)
             self.assertEqual(parsed.checksum_diff, dag.checksum_diff)
+
+
+class ParserDispatcherTests(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.cfg = Config()
+
+    # Can we make a ParserDispatcher without anything blowing up?
+    def test_smoke(self):
+        for arch_ver in TEST_DATA:
+            qzv_fp = TEST_DATA[arch_ver]['qzv_fp']
+            ParserDispatcher(self.cfg, qzv_fp)
+        self.assertTrue(True)
+
+    # TODO NEXT These tests
+    # TODO: Rename to format version?
+    def test_correct_parser(self):
+        for arch_ver in TEST_DATA:
+            qzv_fp = TEST_DATA[arch_ver]['qzv_fp']
+            handler = ParserDispatcher(self.cfg, qzv_fp)
+            self.assertIsInstance(handler.parser,
+                                  TEST_DATA[arch_ver]['parser'])
+
+    def test_correct_parser_type(self):
+        # Confirm we're getting ProvDAGParser v ArtifactParser
+        # Import the types first
+        pass
+
+    def test_parse_with_provdag_parser(self):
+        pass
+
+    def test_parse_with_artifact_parser(self):
+        uuid = TEST_DATA['5']['uuid']
+        qzv_fp = TEST_DATA['5']['qzv_fp']
+        handler = ParserDispatcher(self.cfg, qzv_fp)
+        parser_results = handler.parse(qzv_fp)
+        self.assertIsInstance(parser_results, ParserResults)
+        p_a_uuids = parser_results.parsed_artifact_uuids
+        self.assertIsInstance(p_a_uuids, set)
+        self.assertIsInstance(next(iter(p_a_uuids)), UUID)
+        self.assertEqual(len(parser_results.prov_digraph), 15)
+        self.assertIn(uuid, parser_results.prov_digraph)
+        self.assertIsInstance(
+            parser_results.prov_digraph.nodes[uuid]['node_data'],
+            ProvNode)
+        self.assertEqual(parser_results.provenance_is_valid,
+                         TEST_DATA['5']['prov_is_valid'])
+        self.assertEqual(parser_results.checksum_diff,
+                         TEST_DATA['5']['checksum'])
+
+    def test_no_correct_parser_found_error(self):
+        """
+        Nothing blows up here, it's just not the right kind of filepath
+        e.g. not a zipfile?
+        """
+
+    def test_error_from_one_parser(self):
+        pass
+
+    def test_error_from_two_parsers(self):
+        pass
