+++ provenance_lib/tests/test_zipfile_parser.py
@@ -11,7 +11,6 @@ import zipfile
 from .. import checksum_validator
 from .testing_utilities import is_root_provnode_data
 from .test_parse import TEST_DATA, DATA_DIR
-from ..parse import ParserDispatcher
 from ..util import UUID
 from ..zipfile_parser import (
     ProvNode, Config, _Action, _Citations, _ResultMetadata, ParserResults,
@@ -127,45 +126,6 @@ class ParserVxTests(unittest.TestCase):
             else:
                 parser.parse_prov(Config(), qzv_fp)
                 parser._validate_checksums.assert_called_once()
-
-
-class ParserDispatcherTests(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.cfg = Config()
-
-    # Can we make a ParserDispatcher without anything blowing up?
-    def test_smoke(self):
-        for arch_ver in TEST_DATA:
-            qzv_fp = TEST_DATA[arch_ver]['qzv_fp']
-            ParserDispatcher(self.cfg, qzv_fp)
-        self.assertTrue(True)
-
-    def test_correct_parser(self):
-        for arch_ver in TEST_DATA:
-            qzv_fp = TEST_DATA[arch_ver]['qzv_fp']
-            handler = ParserDispatcher(self.cfg, qzv_fp)
-            self.assertIsInstance(handler.parser,
-                                  TEST_DATA[arch_ver]['parser'])
-
-    def test_parse(self):
-        uuid = TEST_DATA['5']['uuid']
-        qzv_fp = TEST_DATA['5']['qzv_fp']
-        handler = ParserDispatcher(self.cfg, qzv_fp)
-        parser_results = handler.parse(qzv_fp)
-        self.assertIsInstance(parser_results, ParserResults)
-        p_a_uuids = parser_results.parsed_artifact_uuids
-        self.assertIsInstance(p_a_uuids, set)
-        self.assertIsInstance(next(iter(p_a_uuids)), UUID)
-        self.assertEqual(len(parser_results.prov_digraph), 15)
-        self.assertIn(uuid, parser_results.prov_digraph)
-        self.assertIsInstance(
-            parser_results.prov_digraph.nodes[uuid]['node_data'],
-            ProvNode)
-        self.assertEqual(parser_results.provenance_is_valid,
-                         TEST_DATA['5']['prov_is_valid'])
-        self.assertEqual(parser_results.checksum_diff,
-                         TEST_DATA['5']['checksum'])
 
 
 class ArtifactParserTests(unittest.TestCase):
